requirements.txt
Flask==3.0.3
flask-cors==5.0.0
google-genai==0.3.0
python-dotenv==1.0.1
